feat: re did structure
