+++ .github/workflows/data-pipeline.yml
@@ -25,7 +25,7 @@ jobs:
         shell: bash
 
       - name: Run script
-        run: yarn pipeline:run
+        run: yarn pipeline:run:iesg
         shell: bash
 
       - name: Publish to Cloudflare Pages
@@ -34,6 +34,6 @@ jobs:
           apiToken: ${{ secrets.CLOUDFLARE_API_TOKEN }}
           accountId: "c998444368535e9100cd4e83bb76fb41"
           projectName: "iesg"
-          directory: "public"
+          directory: "public/iesg"
           wranglerVersion: "3"
           gitHubToken: ${{ secrets.GITHUB_TOKEN }}
