--- .github/workflows/deploy.yml
@@ -1,25 +1,0 @@
-name: Deploy
-
-on:
-  push:
-    branches: ["main"]
-
-jobs:
-  deploy:
-    runs-on: ubuntu-latest
-    permissions:
-      contents: read
-      deployments: write
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-
-      - name: Publish to Cloudflare Pages
-        uses: cloudflare/pages-action@v1
-        with:
-          apiToken: ${{ secrets.CLOUDFLARE_API_TOKEN }}
-          accountId: "c998444368535e9100cd4e83bb76fb41"
-          projectName: "iesg"
-          directory: "public"
-          wranglerVersion: "3"
-          gitHubToken: ${{ secrets.GITHUB_TOKEN }}
