--- gb/public/styles/index.css
@@ -1,209 +1,0 @@
-body {
-  font-family: 'Courier New', monospace;
-  font-size: 16px;
-  margin-left: auto;
-  margin-right: auto;
-  width: 30%;
-}
-
-p {
-  text-align: left;
-  font-weight: 400;
-  font-style: normal;
-  line-height: 1.9;
-}
-
-h1,
-h2 {
-  text-align: center;
-}
-
-.innerTable {
-  margin: auto;
-  width: 50px;
-  table-layout: auto;
-  border-collapse: collapse;
-  border-style: hidden;
-}
-
-.innerTable th {
-  font-size: small;
-  text-align: center;
-  font-weight: bold;
-  background-color: inherit;
-  padding: 4px;
-}
-
-.innerTable td {
-  font-size: small;
-  text-align: center;
-  background-color: inherit;
-  padding: 4px;
-  font-weight: inherit;
-}
-
-.tableContainer{
-  height: 300px; 
-  overflow: auto; 
-}
-
-table {
-  text-align: left;
-  margin: auto;
-  width: 100%;
-  table-layout: fixed;
-}
-
-td,
-th {
-  border: 1px solid #dddddd;
-  padding: 8px;
-  font-weight: 300;
-  font-style: normal;
-  font-size: small;
-  text-align: center;
-  width: 90px;
-  overflow: hidden;
-}
-
-th {
-  position: sticky;
-  top: 0;
-  background-color: white;
-  cursor: pointer;
-}
-
-.inputContainer {
-  display: flex;
-  flex-direction: column;
-  gap: 5px;
-  font-size: 16px;
-}
-
-input {
-  font-family: 'Courier New', monospace;
-  font-size: 16px;
-}
-
-input::placeholder {
-  font-size: 16px;
-}
-
-input[type="search"]::-webkit-search-cancel-button {
-  -webkit-appearance: none;
-  height: 16px;
-  width: 16px;
-  margin-left: 0.4em;
-  background-image: url("data:image/svg+xml;utf8,<svg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 24 24' fill='%23777'><path d='M19 6.41L17.59 5 12 10.59 6.41 5 5 6.41 10.59 12 5 17.59 6.41 19 12 13.41 17.59 19 19 17.59 13.41 12z'/></svg>");
-  cursor: pointer;
-}
-
-.formItemSelect {
-  display: flex;
-  flex-direction: row;
-  gap: 10px;
-}
-
-.footerContainer {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  margin-top: 20px;
-}
-
-.footerImageContainer {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  gap: 40px;
-}
-
-.footerTextContainer {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-}
-
-.button {
-  border-radius: 8px;
-  border-width: 0;
-  cursor: pointer;
-  display: inline-block;
-  font-size: 16px;
-  font-weight: 500;
-  line-height: 20px;
-  list-style: none;
-  margin: 0;
-  padding: 10px 12px;
-  text-align: center;
-  transition: all 200ms;
-  vertical-align: baseline;
-  white-space: nowrap;
-  user-select: none;
-  -webkit-user-select: none;
-  touch-action: manipulation;
-}
-
-select {
-  font-weight: 400;
-  font-family: 'Courier New', monospace;
-  font-size: 16px;
-}
-
-path {
-  fill: lightgray;
-  stroke: white;
-}
-
-.foreground {
-  fill: none;
-  stroke: white;
-  stroke-width: 1.5px;
-}
-
-.mapContainer {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.legend-item {
-  font-size: 12px;
-}
-
-.mapTitle {
-  font-size: 12px;
-}
-
-.tooltip{
-  font-size: 12px;
-}
-
-.headerSortDown:after,
-.headerSortUp:after {
-  content: " ";
-  position: relative;
-  left: 10px;
-  border: 7px outset transparent;
-}
-
-.headerSortDown:after {
-  top: 10px;
-  border-top-color: black;
-}
-
-.headerSortUp:after {
-  bottom: 10px;
-  border-bottom-color: black;
-}
-
-@media screen and (max-width: 800px) {
-  body {
-    width: 90%;
-    font-size: 14px;
-  }
-  td,
-  th {
-    font-size: smaller;
-  }
-}
