--- pipelines/gb/configuration.mjs
@@ -1,17 +1,0 @@
-export const config = {
-  years: ["2022"],
-  propertyPrices: {
-    fetch: false,
-  },
-  titles: {
-    mimhTitle: "Medianpris / medianlön",
-    indexTitle: "BFI",
-    mainTitle: "Göteborgs Bostadsförmågaindex",
-  },
-  scales: {
-    gothenburgMedianIncomeToMedianHousePrice: 0.45,
-    ownershipRate: 0.25,
-    netMigration: 0.15,
-    overCrowdingRate: 0.15,
-  },
-};
