+++ pipelines/gbpi/configuration.mjs
@@ -1,0 +1,17 @@
+export const config = {
+  years: ["2022"],
+  propertyPrices: {
+    fetch: false,
+  },
+  titles: {
+    mimhTitle: "Medianpris / medianlön",
+    indexTitle: "Göteborgs Boendeprisindex",
+    indexShortTitle: "GBPI",
+  },
+  scales: {
+    gothenburgMedianIncomeToMedianHousePrice: 0.45,
+    ownershipRate: 0.25,
+    netMigration: 0.15,
+    overCrowdingRate: 0.15,
+  },
+};
