--- public/robots.txt
@@ -1,3 +1,0 @@
-# Allow all crawlers to access all content
-User-agent: *
-Disallow:
