--- public/scripts/helpers.mjs
@@ -1,3 +1,0 @@
-export const round = (number) => Math.round(number * 100) / 100;
-
-export const getIndexValue = (raw) => Math.round(round(raw) * 100);
